Move VintageNet.info implementation to module
The VintageNet.info implementation is getting pretty long so rather than
clutter up vintage_net.ex, move it to its own module.
